Overall DDI risk summary text block

Usage

ddi_risk_summary(
  perp,
  cyp_inh = NULL,
  cyp_tdi = NULL,
  cyp_ind = NULL,
  ugt_inh = NULL,
  transporter_inh = NULL,
  d = 1,
  include_induction = TRUE,
  substr = cyp_reference_substrates,
  cyp_kdeg = cyp_turnover
)

Arguments

perp

The perpetrator object.

cyp_inh

CYP inhibition data as data frame. The following fields are expected:

'name' The name of the perpetrator compound.

'cyp' The CYP enzyme as (upper case) character.

'ki' The \(K_i\) in \(\mu M\) as numeric.

'source' Optional source information as character.

cyp_tdi

The CYP TDI data as data frame. The following fields are expected:

'name' The perpetrator compound name as character.

'cyp' The CYP enzyme as character.

'ki' The \(K_I\) in \(\mu M\) as numeric.

'kinact' The \(k_{inact}\) in 1/h as numeric.

'source' Optional source information as character,

cyp_ind

The CYP induction data as data frame. The following fields are expected:

'name' The name of the perpetrator compound as character.

'cyp' The CYP enzyme as (upper case) character.

'emax' The \(E_{max}\), i.e., the maximum induction effect determined in vitro as numeric.

'ec50' The \(EC_{50}\) in \(\mu M\) as numeric.

'maxc' The maximal concentration in \(\mu M\) tested in the in vitro assay as numeric.

'source' Optional source information as character.

ugt_inh

UGT inhibition data as data frame, The following fields are expected:

'name' The name of the perpetrator compound.

'ugt' The UGT enzyme as (upper case) character.

'ic50' The \(IC_{50}\) in \(\mu M\) as numeric.

'source' Optional source information as character.

transporter_inh

Transporter inhibition data as data frame. The following fields are expected:

'name' The perpetrator compound name.

'cyp' The UGT enzyme as (upper case) character.

'ic50' The \(IC_{50}\) of the inhibition in μM.

'source' Optional source information as character.

d

Scaling factor, defaults to 1.

include_induction

Switch to define whether induction effects should be included in the calculation (C-terms as per the FDA guideline)

substr

The CYP probe substrates to be used as data frame, defaults to cyp_reference_substrates. The data frame is expected to have the following fields:

'cyp' The CYP enzyme as (upper case) character.

'substrate' The substrate name as character.

'fgut' The fraction of the drug escaping gut metabolism.

'fm' The fraction of the drug that undergoes hepatic metabolism.

'fmcyp' The fraction metabolized by the respective CYP enzyme.

cyp_kdeg

The CYP turnover data as data frame. Defaults to the built-in reference data, cyp_turnover.

Value

Markdown-formatted text output.

Examples

ddi_risk_summary(examplinib_compounds, examplinib_cyp_inhibition_data,
  examplinib_cyp_tdi_data, examplinib_cyp_induction_data,
  examplinib_ugt_inhibition_data, examplinib_transporter_inhibition_data)
#> [1] "* examplinib has a clinical risk for direct inhibition of CYP2C9 and CYP2C19 (basic modeling)\n* examplinib has a clinical risk for time-dependent inhibition of CYP3A4\n* examplinib has a clinical risk for induction of CYP3A4 (fold-change method)\n* examplinib has a clinical risk for induction of CYP3A4 (basic kinetic method)\n* M1 has a clinical risk for induction of CYP2B6 and CYP3A4 (fold-change method)\n* M1 has a clinical risk for induction of CYP2B6 and CYP3A4 (basic kinetic method)\n* based on mechanistic-static modeling (S-warfarin and omeprazole), examplinib has a clinical risk for inhibition of CYP2C9 and CYP2C19\n* based on mechanistic-static modeling (midazolam), examplinib has a clinical risk for induction of CYP3A4\n* examplinib has a clinical risk for inhibition of UGT1A1, UGT1A3, UGT1A4, UGT1A6, UGT1A9, UGT2B7, UGT2B15 and UGT2B17\n* M1 has a clinical risk for inhibition of UGT1A1 and UGT2B17\n* examplinib has a clinical risk for inhibition of P-gp (intestinal), P-gp (hepatic), BCRP (intestinal), BCRP (hepatic), MATE1 and MATE2k"
ddi_risk_summary(examplinib_parent, examplinib_cyp_inhibition_data)
#> [1] "* examplinib has a clinical risk for direct inhibition of CYP2C9 and CYP2C19 (basic modeling)\n* based on mechanistic-static modeling (S-warfarin and omeprazole), examplinib has a clinical risk for inhibition of CYP2C9 and CYP2C19"