Basic kinetic CYP induction risk

Basic kinetic CYP induction risk

Usage

kinetic_cyp_induction_risk(perp, cyp_ind, d = 1)

Arguments

perp

The perpetrator object or a list thereof.

cyp_ind

The CYP induction data as data frame. The following fields are expected:

'name' The name of the perpetrator compound as character.

'cyp' The CYP enzyme as (upper case) character.

'emax' The \(E_{max}\), i.e., the maximum induction effect determined in vitro as numeric.

'ec50' The \(EC_{50}\) in \(\mu M\) as numeric.

'maxc' The maximal concentration in \(\mu M\) tested in the in vitro assay as numeric.

'source' Optional source information as character.

d

Scaling factor, defaults to 1.

Value

A data frame.

Details

Basic kinetic modeling of the CYP induction risk considers:

$$R = \frac {1}{1 + d * \frac {E_{max}*10*C_{max,ss,u}}{EC_{50,u} + 10 * C_{max,ss,u}}}$$

\(d\) is a scaling factor with a standard value of 1. A different value can be used if warranted by prior experience with the experimental conditions.

\(R \le 0.8\) suggest a relevant in vivo CYP induction potential.

Examples

kinetic_cyp_induction_risk(examplinib_parent, examplinib_cyp_induction_data)
#>       cyp emax ec50 maxc    source         r risk
#> 1  CYP1A2 1.00   NA    5 study 007        NA   NA
#> 2  CYP2B6 1.00   NA    5 study 007        NA   NA
#> 3  CYP2C8   NA   NA   NA                  NA   NA
#> 4  CYP2C9   NA   NA   NA                  NA   NA
#> 5 CYP2C19   NA   NA   NA                  NA   NA
#> 6  CYP2D6   NA   NA   NA                  NA   NA
#> 7  CYP3A4 7.35 1.64    3 study 007 0.2134856 TRUE