Make a perpetrator object from compound data

This function creates a perpetrator object from key compound data. Note that the 'source' field is empty. If you want to include source information, you can either use new_perpetrator() which takes as input a data that may include source information, or you can create a perpetrator object from a text input string using read_perpetrators().

Usage

make_perpetrator(
  name,
  dose,
  imaxss,
  mw,
  type = "parent",
  oral = TRUE,
  fu = 1,
  fumic = 1,
  rb = 1,
  fa = 1,
  fg = 1,
  ka = 0.1,
  solubility = Inf
)

Arguments

name

The compound name as character.

dose

The clinical dose in mg.

imaxss

The (total) steady-state Cmax in ng/ml.

mw

The molar weight in g/mol.

type

The compound type as character. Must be either 'parent' or metabolite'.

oral

Oral administration as Boolean. Defaults to TRUE.

fu

The fraction unbound as numeric. Defaults to 1.

fumic

The fraction unbound in the microsomes. Defaults to 1.

rb

The blood-to-plasma concentration ratio. Defaults to 1.

fa

The fraction absorbed. Defaults to 1.

fg

The fraction escaping gut metabolism. Defaults to 1.

ka

The absorption rate constant in /ml. Defaults to 0.1 /ml.

solubility

The aqueus solubility in mg/l. Defaults to Inf,

Value

A perpetrator object.

See also

Examples

make_perpetrator("test", 100, 1000, 500)
#> ----- DDI perpetrator object -----
#> name         test             
#> oral         TRUE             
#> mw           500              
#> dose         100              
#> imaxss       1000             
#> fu           1                
#> fumic        1                
#> rb           1                
#> fa           1                
#> fg           1                
#> ka           0.1              
#> solubility   Inf