Summary statement for mechanistic-static modeling of CYP modulation

Summary statement for mechanistic-static modeling of CYP modulation

Usage

# S3 method for class 'perpetrator'
mech_stat_cyp_risk_summary(
  perp,
  cyp_inh,
  cyp_ind = NULL,
  cyp_tdi = NULL,
  d = 1,
  include_induction = TRUE,
  substr = cyp_reference_substrates,
  cyp_kdeg = cyp_turnover
)

Arguments

perp

The perpetrator object.

cyp_inh

CYP inhibition data as data frame. The following fields are expected:

'name' The name of the perpetrator compound.

'cyp' The CYP enzyme as (upper case) character.

'ki' The \(K_i\) in \(\mu M\) as numeric.

'source' Optional source information as character.

cyp_ind

The CYP induction data as data frame. The following fields are expected:

'name' The name of the perpetrator compound as character.

'cyp' The CYP enzyme as (upper case) character.

'emax' The \(E_{max}\), i.e., the maximum induction effect determined in vitro as numeric.

'ec50' The \(EC_{50}\) in \(\mu M\) as numeric.

'maxc' The maximal concentration in \(\mu M\) tested in the in vitro assay as numeric.

'source' Optional source information as character.

cyp_tdi

The CYP TDI data as data frame. The following fields are expected:

'name' The perpetrator compound name as character.

'cyp' The CYP enzyme as character.

'ki' The \(K_I\) in \(\mu M\) as numeric.

'kinact' The \(k_{inact}\) in 1/h as numeric.

'source' Optional source information as character,

d

Scaling factor, defaults to 1.

include_induction

Switch to define whether induction effects should be included in the calculation (C-terms as per the FDA guideline)

substr

The CYP probe substrates to be used as data frame, defaults to cyp_reference_substrates. The data frame is expected to have the following fields:

'cyp' The CYP enzyme as (upper case) character.

'substrate' The substrate name as character.

'fgut' The fraction of the drug escaping gut metabolism.

'fm' The fraction of the drug that undergoes hepatic metabolism.

'fmcyp' The fraction metabolized by the respective CYP enzyme.

cyp_kdeg

The CYP turnover data as data frame. Defaults to the built-in reference data, cyp_turnover.

Value

Markdown-formatted text output.