Perpetrator object constructor

Convert a data frame into a perpetrator object.

Usage

new_perpetrator(df)

Arguments

df

A data frame to be converted into a perpetrator object. See 'Details' for the expected format and fields.

Value

A perpetrator object.

Details

The input is a data frame with the columns 'param', 'value' and source'.

Rows with the following parameters ('param') are expected in the input:

PARAM	REQUIREMENT	DESCRIPTION

name	required	the name of the compound

oral	required	logical value (TRUE or FALSE) to indicate whether the compound is administered orally

mw	required	molar weight in g/mol

dose	required for parent compounds	clinical dose in mg

imaxss	required	total Cmax in ng/ml

fu	optional	fraction unbound, default is 1

fumic	optional	microsomal unbound fraction, dafault is 1

rb	optional	blood-to-plasma concentration ratio, default is 1

fa	optional	fraction absorbed, default is 1

fg	optional	fraction escaping gut metabolism, default is 1

ka	optional	absorption rate constant, default is 0.1 /min

solubility	optional	solubility of the compound in mg/l, defaults to Inf

The following is an example for a valid input data frame:

        param      value        source
1        name examplinib
2        oral       TRUE
3          mw      492.6
4        dose        450 clinical dose
5      imaxss       3530     study 001
6          fu      0.023     study 002
7       fumic          1       default
8          rb          1     study 003
9          fa       0.81     study 003
10         fg          1       default
11         ka    0.00267       unknown
12 solubility        Inf       default

See also