Table of perpetrator drug properties

This function generates a markdown-formatted table of the key properties of a perpetrator object.

Usage

property_table(obj)

Arguments

obj

The perpetrator object or a list of perpetrator objects.

Value

A markdown-formatted table.

Examples

property_table(examplinib_parent)
#> 
#> 
#> Table: Compound parameters for examplinib
#> 
#> |parameter            |value   |source        |
#> |:--------------------|:-------|:-------------|
#> |oral                 |TRUE    |              |
#> |$MW$ (g/mol)         |492.6   |              |
#> |$dose$ (mg)          |450     |clinical dose |
#> |$C_{max,ss}$ (ng/ml) |3530    |study 001     |
#> |$f_u$                |0.023   |study 002     |
#> |$f_{u,mic}$          |1       |default       |
#> |$R_B$                |1       |study 003     |
#> |$F_a$                |0.81    |study 003     |
#> |$F_g$                |1       |default       |
#> |$k_a$ (1/min)        |0.00267 |unknown       |
#> |$solubility$ (mg/l)  |Inf     |default       |
property_table(examplinib_compounds)
#> 
#> 
#> Table: Compound parameters for examplinib
#> 
#> |parameter            |value   |source        |
#> |:--------------------|:-------|:-------------|
#> |oral                 |TRUE    |              |
#> |$MW$ (g/mol)         |492.6   |              |
#> |$dose$ (mg)          |450     |clinical dose |
#> |$C_{max,ss}$ (ng/ml) |3530    |study 001     |
#> |$f_u$                |0.023   |study 002     |
#> |$f_{u,mic}$          |1       |default       |
#> |$R_B$                |1       |study 003     |
#> |$F_a$                |0.81    |study 003     |
#> |$F_g$                |1       |default       |
#> |$k_a$ (1/min)        |0.00267 |unknown       |
#> |$solubility$ (mg/l)  |Inf     |default       |
#> 
#> 
#> Table: Compound parameters for M1
#> 
#> |parameter            |value  |source    |
#> |:--------------------|:------|:---------|
#> |oral                 |FALSE  |          |
#> |$MW$ (g/mol)         |506.56 |          |
#> |$dose$ (mg)          |NA     |          |
#> |$C_{max,ss}$ (ng/ml) |1038   |study 001 |
#> |$f_u$                |0.012  |study 002 |
#> |$f_{u,mic}$          |1      |default   |
#> |$R_B$                |1      |study 002 |
#> |$solubility$ (mg/l)  |Inf    |default   |