Table of the basic static CYP induction risk

Table of the basic static CYP induction risk

Usage

static_cyp_induction_risk_table(perp, cyp_ind, na.rm = F, show_dose = FALSE)

Arguments

perp

The perpetrator object or a list thereof.

cyp_ind

The CYP induction data as data frame. The following fields are expected:

'name' The name of the perpetrator compound as character.

'cyp' The CYP enzyme as (upper case) character.

'emax' The \(E_{max}\), i.e., the maximum induction effect determined in vitro as numeric.

'ec50' The \(EC_{50}\) in \(\mu M\) as numeric.

'maxc' The maximal concentration in \(\mu M\) tested in the in vitro assay as numeric.

'source' Optional source information as character.

na.rm

Switch to define whether rows with lacking \(E_{max}\) data are removed from the output. Defaults to FALSE.

show_dose

Show perpetrator dose in table title, defaults to FALSE.

Value

A markdown-formatted table.

See also

Examples

static_cyp_induction_risk_table(examplinib_parent,
  examplinib_cyp_induction_data)
#> 
#> 
#> Table: Risk for hepatic CYP induction by examplinib, basic mRNA fold-change method
#> 
#> |CYP     | $E_{max}$| $max c$ ($\mu M$)|source    | $max c/C_{max,ss,u}$|risk |notes                         |
#> |:-------|---------:|-----------------:|:---------|--------------------:|:----|:-----------------------------|
#> |CYP1A2  |      1.00|                 5|study 007 |                 30.3|No   |Risk cannot be fully excluded |
#> |CYP2B6  |      1.00|                 5|study 007 |                 30.3|No   |Risk cannot be fully excluded |
#> |CYP2C8  |        NA|                NA|          |                   NA|     |                              |
#> |CYP2C9  |        NA|                NA|          |                   NA|     |                              |
#> |CYP2C19 |        NA|                NA|          |                   NA|     |                              |
#> |CYP2D6  |        NA|                NA|          |                   NA|     |                              |
#> |CYP3A4  |      7.35|                 3|study 007 |                 18.2|Yes  |                              |
static_cyp_induction_risk_table(examplinib_compounds,
  examplinib_cyp_induction_data)
#> 
#> 
#> Table: Risk for hepatic CYP induction by examplinib, basic mRNA fold-change method
#> 
#> |CYP     | $E_{max}$| $max c$ ($\mu M$)|source    | $max c/C_{max,ss,u}$|risk |notes                         |
#> |:-------|---------:|-----------------:|:---------|--------------------:|:----|:-----------------------------|
#> |CYP1A2  |      1.00|                 5|study 007 |                 30.3|No   |Risk cannot be fully excluded |
#> |CYP2B6  |      1.00|                 5|study 007 |                 30.3|No   |Risk cannot be fully excluded |
#> |CYP2C8  |        NA|                NA|          |                   NA|     |                              |
#> |CYP2C9  |        NA|                NA|          |                   NA|     |                              |
#> |CYP2C19 |        NA|                NA|          |                   NA|     |                              |
#> |CYP2D6  |        NA|                NA|          |                   NA|     |                              |
#> |CYP3A4  |      7.35|                 3|study 007 |                 18.2|Yes  |                              |
#> 
#> 
#> Table: Risk for hepatic CYP induction by M1, basic mRNA fold-change method
#> 
#> |CYP     | $E_{max}$| $max c$ ($\mu M$)|source    | $max c/C_{max,ss,u}$|risk |notes |
#> |:-------|---------:|-----------------:|:---------|--------------------:|:----|:-----|
#> |CYP1A2  |      1.00|                 5|study 007 |                203.3|No   |      |
#> |CYP2B6  |      6.98|                 5|study 007 |                203.3|Yes  |      |
#> |CYP2C8  |        NA|                NA|          |                   NA|     |      |
#> |CYP2C9  |        NA|                NA|          |                   NA|     |      |
#> |CYP2C19 |        NA|                NA|          |                   NA|     |      |
#> |CYP2D6  |        NA|                NA|          |                   NA|     |      |
#> |CYP3A4  |     22.70|                 5|study 007 |                203.3|Yes  |      |
static_cyp_induction_risk_table(examplinib_compounds,
  examplinib_cyp_induction_data, show_dose = TRUE)
#> 
#> 
#> Table: Risk for hepatic CYP induction by examplinib (450 mg), basic mRNA fold-change method
#> 
#> |CYP     | $E_{max}$| $max c$ ($\mu M$)|source    | $max c/C_{max,ss,u}$|risk |notes                         |
#> |:-------|---------:|-----------------:|:---------|--------------------:|:----|:-----------------------------|
#> |CYP1A2  |      1.00|                 5|study 007 |                 30.3|No   |Risk cannot be fully excluded |
#> |CYP2B6  |      1.00|                 5|study 007 |                 30.3|No   |Risk cannot be fully excluded |
#> |CYP2C8  |        NA|                NA|          |                   NA|     |                              |
#> |CYP2C9  |        NA|                NA|          |                   NA|     |                              |
#> |CYP2C19 |        NA|                NA|          |                   NA|     |                              |
#> |CYP2D6  |        NA|                NA|          |                   NA|     |                              |
#> |CYP3A4  |      7.35|                 3|study 007 |                 18.2|Yes  |                              |
#> 
#> 
#> Table: Risk for hepatic CYP induction by M1, basic mRNA fold-change method
#> 
#> |CYP     | $E_{max}$| $max c$ ($\mu M$)|source    | $max c/C_{max,ss,u}$|risk |notes |
#> |:-------|---------:|-----------------:|:---------|--------------------:|:----|:-----|
#> |CYP1A2  |      1.00|                 5|study 007 |                203.3|No   |      |
#> |CYP2B6  |      6.98|                 5|study 007 |                203.3|Yes  |      |
#> |CYP2C8  |        NA|                NA|          |                   NA|     |      |
#> |CYP2C9  |        NA|                NA|          |                   NA|     |      |
#> |CYP2C19 |        NA|                NA|          |                   NA|     |      |
#> |CYP2D6  |        NA|                NA|          |                   NA|     |      |
#> |CYP3A4  |     22.70|                 5|study 007 |                203.3|Yes  |      |