Basic CYP inhibition risk

This function evaluates the clinical risk for direct (reversible) CYP inhibition according to the basic model defined in the relevant regulatory guideline.

Usage

basic_cyp_inhibition_risk(perp, cyp_inh)

Arguments

perp

The perpetrator object.

cyp_inh

CYP inhibition data as data frame. The following fields are expected:

'name' The name of the perpetrator compound.

'cyp' The CYP enzyme as (upper case) character.

'ki' The \(K_i\) in \(\mu M\) as numeric.

'source' Optional source information as character.

Value

A data frame.

Details

For the basic modeling of direct (reversible) CYP enzyme inhibition, the ratio of the relevant inhibitor concentration to the \(K_i\) of the respective CYP enzyme is considered, i.e., \(R\) for hepatic enzymes and \(R_{gut}\) for intestinal enzymes (refer to Section 2.1.2.1 of the ICH M12 guidance document).

Liver

$$R=\frac{C_{max,ss,u}}{K_{i,u}}$$

\(R\) values > 0.02, i.e., maximal unbound perpetrator concentrations 50-fold over \(K_i\) are considered to indicate a potential clinical CYP inhibition risk using this method.

Intestine

$$R_{gut}=\frac{I_{gut}}{K_{i,u}}$$

where

$$I_{gut}=\frac{Dose}{250\ mg}$$

\(R\) values > 10 are considered to indicate a clinical risk for intestinal CYP3A inhibition.

In the output, the columns risk_hep and risk_intest indicate whether the regulatory threshold is reached for the respective enzyme.

See also

Examples

basic_cyp_inhibition_risk(examplinib_parent, examplinib_cyp_inhibition_data)
#>       cyp    ki   kiu      r risk_hep    r_gut risk_intest
#> 1  CYP1A2    NA    NA     NA       NA       NA          NA
#> 2  CYP2B6    NA    NA     NA       NA       NA          NA
#> 3  CYP2C8 11.00 11.00 0.0150    FALSE       NA          NA
#> 4  CYP2C9  0.60  0.60 0.2747     TRUE       NA          NA
#> 5 CYP2C19  0.25  0.25 0.6593     TRUE       NA          NA
#> 6  CYP2D6    NA    NA     NA       NA       NA          NA
#> 7  CYP3A4 12.50 12.50 0.0132    FALSE 292.3264        TRUE