Basic CYP inhibition risk

This function evaluates the clinical risk for direct (reversible) CYP inhibition according to the basic model defined in the relevant regulatory guideines.

Usage

basic_cyp_inhibition_risk(perp, cyp_inh)

Arguments

perp

The perpetrator object.

cyp_inh

CYP inhibition data as data frame.

Value

A data frame.

Details

For the basic modeling of direct (reversible) CYP enzyme inhibition, the ratios of the relevant inhibitor concentration to the \(K_i\) are considered, i.e., \(R_1\) for hepatic enzymes and \(R_{1,gut}\) for intestinal enzymes (refer to fig. 1 of the the FDA guidance). A threshold of 1.02 applies.

Note that section 5.3.3.1 of the EMA DDI guidance defines the ratio as \(R_1=\frac{I_{max,ss,u}}{K_{i,u}}\) and the threshold as 0.02.

Liver

$$R_1=1+\frac{I_{max,ss,u}}{K_{i,u}}$$

Gut wall

$$R_{1,gut}=1+\frac{I_{gut}}{K_{i,u}}$$

Refer to the documentation to the key_concentrations() function for details on the calculation of \(I_{max,ss,u}\) and \(I_{gut}\).

See also

Examples

basic_cyp_inhibition_risk(examplinib_parent, examplinib_cyp_inhibition_data)
#>     param   ki  kiu    r1 risk_hep r1gut risk_intest
#> 1  CYP1A2   NA   NA    NA       NA    NA          NA
#> 2  CYP2B6   NA   NA    NA       NA    NA          NA
#> 3  CYP2C8 11.0 11.0 1.015    FALSE    NA          NA
#> 4  CYP2C9 13.5 13.5 1.012    FALSE    NA          NA
#> 5 CYP2C19 15.0 15.0 1.011    FALSE    NA          NA
#> 6  CYP2D6   NA   NA    NA       NA    NA          NA
#> 7  CYP3A4 12.5 12.5 1.013    FALSE 293.3        TRUE