UGT inhibition risk

This function evaluates the clinical risk for reversible inhibition of UGT enzymes according to the relevant regulatory guideline.

Usage

basic_ugt_inhibition_risk(perp, ugt_inh)

Arguments

perp

The perpetrator object.

ugt_inh

UGT inhibition data as data frame, The following fields are expected:

'name' The name of the perpetrator compound.

'ugt' The UGT enzyme as (upper case) character.

'ic50' The \(IC_{50}\) in \(\mu M\) as numeric.

'source' Optional source information as character.

Value

A markdown-formatted table.

Details

This function assumes that the UGT inhibition data is provided as \(IC_{50}\). According to Cheng, Prusoff 1973), \(K_i\) can be assumed to be \(IC_{50}/2\) at the experimental conditions commonly used in the in vitro inhibition studies where substrate concentrations are close to \(K_M\).

The relevant metric for basic modeling of the UGT inhibition risk is \(R=C_{max,ss,u}/K_{i,u}\)

Refer to Section 2.1.2.1 of the ICH M12 guidance document) for details.

\(R>0.02\) are considered to indicate a potential UGT inhibition risk.

See also

Examples

basic_ugt_inhibition_risk(examplinib_parent, examplinib_ugt_inhibition_data)
#>       ugt  kiu          r risk
#> 1  UGT1A1 7.50 0.02197591 TRUE
#> 2  UGT1A3 7.50 0.02197591 TRUE
#> 3  UGT1A4 7.50 0.02197591 TRUE
#> 4  UGT1A6 7.50 0.02197591 TRUE
#> 5  UGT1A9 1.90 0.08674701 TRUE
#> 6  UGT2B7 7.50 0.02197591 TRUE
#> 7 UGT2B15 7.50 0.02197591 TRUE
#> 8 UGT2B17 3.05 0.05403912 TRUE