CYP perpetration risk as per mechanistic-static modeling

CYP perpetration risk as per mechanistic-static modeling

Usage

mech_stat_cyp_risk(
  perp,
  cyp_inh,
  cyp_ind,
  cyp_tdi = NULL,
  d = 1,
  include_induction = TRUE,
  substr = cyp_reference_substrates,
  cyp_kdeg = cyp_turnover
)

Arguments

perp

The perpetrator object.

cyp_inh

CYP inhibition data as data frame. The following fields are expected:

'name' The name of the perpetrator compound.

'cyp' The CYP enzyme as (upper case) character.

'ki' The \(K_i\) in \(\mu M\) as numeric.

'source' Optional source information as character.

cyp_ind

The CYP induction data as data frame. The following fields are expected:

'name' The name of the perpetrator compound as character.

'cyp' The CYP enzyme as (upper case) character.

'emax' The \(E_{max}\), i.e., the maximum induction effect determined in vitro as numeric.

'ec50' The \(EC_{50}\) in \(\mu M\) as numeric.

'maxc' The maximal concentration in \(\mu M\) tested in the in vitro assay as numeric.

'source' Optional source information as character.

cyp_tdi

The CYP TDI data as data frame. The following fields are expected:

'name' The perpetrator compound name as character.

'cyp' The CYP enzyme as character.

'ki' The \(K_I\) in \(\mu M\) as numeric.

'kinact' The \(k_{inact}\) in 1/h as numeric.

'source' Optional source information as character,

d

Scaling factor, defaults to 1.

include_induction

Switch to define whether induction effects should be included in the calculation (C-terms as per the FDA guideline)

substr

The CYP probe substrates to be used as data frame, defaults to cyp_reference_substrates. The data frame is expected to have the following fields:

'cyp' The CYP enzyme as (upper case) character.

'substrate' The substrate name as character.

'fgut' The fraction of the drug escaping gut metabolism.

'fm' The fraction of the drug that undergoes hepatic metabolism.

'fmcyp' The fraction metabolized by the respective CYP enzyme.

cyp_kdeg

The CYP turnover data as data frame. Defaults to the built-in reference data, cyp_turnover.

Value

A data frame.

Details

In this approach, AUC ratios for CYP-specific probe substrates are calculated based on their known intestinal and hepatic metabolism. Direct (competitive) and time-dependent inhibition terms, as well as enzyme induction terms are considered. For details, refer to section 7.5.1.2 of the ICH M12 guideline.

$$AUCR = \frac{1}{A_g*B_g*C_g* \left(1-F_g \right)+F_g} * \frac{1}{A_h*B_h*C_h*f_m+\left(1-f_m\right)}$$

The individual terms are:

Reversible inhibition

$$A_g = \frac{1}{1+\frac{I_g}{K_i}}$$

$$A_h = \frac{1}{1+\frac{I_h}{K_i}}$$

Time-dependent inhibition

$$B_g = \frac{k_{deg,g}}{k_{deg,g} + \frac{I_g*k_{inact}}{I_g+K_I}}$$

$$B_h = \frac{k_{deg,h}}{k_{deg,h} + \frac{I_h*k_{inact}}{I_h+K_I}}$$

Induction

$$C_g = 1 + \frac{d*E_{max}*I_g}{I_g+EC_{50}}$$

$$C_h = 1 + \frac{d*E_{max}*I_h}{I_h+EC_{50}}$$

with the hepatic inlet concentration \(I_h=I_{max,inlet,u}\) and the intestinal concentration \(I_g=I_{enteric,u}\), see key_concentrations().

\(d\) is a scaling factor for the CYP induction term with a standard value of 1. A different value can be used if warranted by prior experience with the experimental setup.

See also

Examples

mech_stat_cyp_risk(examplinib_parent, examplinib_cyp_inhibition_data,
  examplinib_cyp_induction_data)
#>       cyp   substrate   kiu fgut   fm fmcyp         Ag        Ah Bg Bh       Cg
#> 1  CYP1A2  tizanidine    NA 1.00 0.95  0.98 1.00000000 1.0000000  1  1 1.000000
#> 2  CYP2B6        <NA>    NA   NA   NA    NA 1.00000000 1.0000000  1  1 1.000000
#> 3  CYP2C8 repaglinide 11.00 1.00 1.00  0.61 0.62551298 0.9827625  1  1 1.000000
#> 4  CYP2C9  S-warfarin  0.60 1.00 1.00  0.91 0.08350072 0.7566792  1  1 1.000000
#> 5 CYP2C19  omeprazole  0.25 1.00 1.00  0.87 0.03657341 0.5644126  1  1 1.000000
#> 6  CYP2D6 desipramine    NA 1.00 1.00  0.85 1.00000000 1.0000000  1  1 1.000000
#> 7  CYP3A4   midazolam 12.50 0.57 0.96  1.00 0.65494520 0.9847995  1  1 6.884569
#>         Ch      aucr  risk
#> 1 1.000000 1.0000000 FALSE
#> 2 1.000000        NA    NA
#> 3 1.000000 1.0106266 FALSE
#> 4 1.000000 1.2843927  TRUE
#> 5 1.000000 1.6102050  TRUE
#> 6 1.000000 1.0000000 FALSE
#> 7 1.773674 0.2321612  TRUE
mech_stat_cyp_risk(examplinib_parent, examplinib_cyp_inhibition_data,
  examplinib_cyp_induction_data, examplinib_cyp_tdi_data)
#>       cyp   substrate   kiu fgut   fm fmcyp         Ag        Ah        Bg
#> 1  CYP1A2  tizanidine    NA 1.00 0.95  0.98 1.00000000 1.0000000 1.0000000
#> 2  CYP2B6        <NA>    NA   NA   NA    NA 1.00000000 1.0000000 1.0000000
#> 3  CYP2C8 repaglinide 11.00 1.00 1.00  0.61 0.62551298 0.9827625 1.0000000
#> 4  CYP2C9  S-warfarin  0.60 1.00 1.00  0.91 0.08350072 0.7566792 1.0000000
#> 5 CYP2C19  omeprazole  0.25 1.00 1.00  0.87 0.03657341 0.5644126 1.0000000
#> 6  CYP2D6 desipramine    NA 1.00 1.00  0.85 1.00000000 1.0000000 1.0000000
#> 7  CYP3A4   midazolam 12.50 0.57 0.96  1.00 0.65494520 0.9847995 0.4348241
#>          Bh       Cg       Ch      aucr  risk
#> 1 1.0000000 1.000000 1.000000 1.0000000 FALSE
#> 2 1.0000000 1.000000 1.000000        NA    NA
#> 3 1.0000000 1.000000 1.000000 1.0106266 FALSE
#> 4 1.0000000 1.000000 1.000000 1.2843927  TRUE
#> 5 1.0000000 1.000000 1.000000 1.6102050  TRUE
#> 6 1.0000000 1.000000 1.000000 1.0000000 FALSE
#> 7 0.4757909 6.884569 1.773674 0.8446594 FALSE
mech_stat_cyp_risk(examplinib_parent, examplinib_cyp_inhibition_data,
  examplinib_cyp_induction_data, examplinib_cyp_tdi_data,
  include_induction = FALSE)
#>       cyp   substrate   kiu fgut   fm fmcyp         Ag        Ah        Bg
#> 1  CYP1A2  tizanidine    NA 1.00 0.95  0.98 1.00000000 1.0000000 1.0000000
#> 2  CYP2B6        <NA>    NA   NA   NA    NA 1.00000000 1.0000000 1.0000000
#> 3  CYP2C8 repaglinide 11.00 1.00 1.00  0.61 0.62551298 0.9827625 1.0000000
#> 4  CYP2C9  S-warfarin  0.60 1.00 1.00  0.91 0.08350072 0.7566792 1.0000000
#> 5 CYP2C19  omeprazole  0.25 1.00 1.00  0.87 0.03657341 0.5644126 1.0000000
#> 6  CYP2D6 desipramine    NA 1.00 1.00  0.85 1.00000000 1.0000000 1.0000000
#> 7  CYP3A4   midazolam 12.50 0.57 0.96  1.00 0.65494520 0.9847995 0.4348241
#>          Bh Cg Ch     aucr  risk
#> 1 1.0000000  1  1 1.000000 FALSE
#> 2 1.0000000  1  1       NA    NA
#> 3 1.0000000  1  1 1.010627 FALSE
#> 4 1.0000000  1  1 1.284393  TRUE
#> 5 1.0000000  1  1 1.610205  TRUE
#> 6 1.0000000  1  1 1.000000 FALSE
#> 7 0.4757909  1  1 2.948311  TRUE
mech_stat_cyp_risk(examplinib_parent, examplinib_cyp_inhibition_data, NULL)
#>       cyp   substrate   kiu fgut   fm fmcyp         Ag        Ah Bg Bh Cg Ch
#> 1  CYP1A2  tizanidine    NA 1.00 0.95  0.98 1.00000000 1.0000000  1  1  1  1
#> 2  CYP2B6        <NA>    NA   NA   NA    NA 1.00000000 1.0000000  1  1  1  1
#> 3  CYP2C8 repaglinide 11.00 1.00 1.00  0.61 0.62551298 0.9827625  1  1  1  1
#> 4  CYP2C9  S-warfarin  0.60 1.00 1.00  0.91 0.08350072 0.7566792  1  1  1  1
#> 5 CYP2C19  omeprazole  0.25 1.00 1.00  0.87 0.03657341 0.5644126  1  1  1  1
#> 6  CYP2D6 desipramine    NA 1.00 1.00  0.85 1.00000000 1.0000000  1  1  1  1
#> 7  CYP3A4   midazolam 12.50 0.57 0.96  1.00 0.65494520 0.9847995  1  1  1  1
#>       aucr  risk
#> 1 1.000000 FALSE
#> 2       NA    NA
#> 3 1.010627 FALSE
#> 4 1.284393  TRUE
#> 5 1.610205  TRUE
#> 6 1.000000 FALSE
#> 7 1.191612 FALSE