CYP perpetration risk as per mechanistic-static modeling

CYP perpetration risk as per mechanistic-static modeling

Usage

mech_stat_cyp_risk(
  perp,
  cyp_inh,
  cyp_ind,
  include_induction = T,
  substr = cyp_reference_substrates
)

Arguments

perp

The perpetrator object.

cyp_inh

CYP inhibiton data as data frame.

cyp_ind

CYP induction data as data frame.

include_induction

Boolean value to define whether induction should be included in the calculation (C-terms as per the FDA guideline)

substr

The CYP reference substrates to be used as data frame.

Value

A data frame.

Details

In this approach, AUC ratios for probe substrates are calculated based on their known intestinal and hepatic metabolism. Both direct (competitive) and time-dependent inhibition are considered.

The below formula given by the FDA guideline (refer to FDA 2020, Fig. 7) also includes intestinal and hepatic enzyme induction terms (\(C_g\) and \(C_h\), respectively). At the same time, the guideline states that both inhibition and induction should be considered separately.

$$AUCR = \frac{1}{A_g*B_g*C_g* \left(1-F_g \right)+F_g} * \frac{1}{A_h*B_h*C_h*f_m+\left(1-f_m\right)}$$

The individual terms are:

Reversible inhibition

$$A_g = \frac{1}{1+\frac{I_g}{K_i}}$$

$$A_h = \frac{1}{1+\frac{I_h}{K_i}}$$

Time-dependent inhibition

$$B_g = \frac{k_{deg,g}}{k_{deg,g} + \frac{I_g*k_{inact}}{I_g+K_I}}$$

$$B_h = \frac{k_{deg,h}}{k_{deg,h} + \frac{I_h*k_{inact}}{I_h+K_I}}$$

Induction

$$C_g = 1 + \frac{d*E_{max}*I_g}{I_g+EC_{50}}$$

$$C_h = 1 + \frac{d*E_{max}*I_h}{I_h+EC_{50}}$$

with the hepatic inlet concentration \(I_h=I_{max,inlet,u}\) and the intestinal concentration \(I_g=I_{enteric,u}\), see key_concentrations().

\(d\) is an induction scaling factor (assumed to be 1, but can be adjusted).

Examples

mech_stat_cyp_risk(examplinib_parent, examplinib_cyp_inhibition_data, examplinib_cyp_induction_data)
#>       cyp   substrate  kiu fgut   fm fmcyp        Ag        Ah      Cg       Ch
#> 1  CYP1A2  tizanidine   NA 1.00 0.95  0.98        NA        NA 1.00000 1.000000
#> 2  CYP2B6        <NA>   NA   NA   NA    NA        NA        NA 1.00000 1.000000
#> 3  CYP2C8 repaglinide 11.0 1.00 1.00  0.61 0.9864172 0.9827625 1.00000 1.000000
#> 4  CYP2C9  S-warfarin 13.5 1.00 1.00  0.91 0.9889046 0.9859096 1.00000 1.000000
#> 5 CYP2C19  omeprazole 15.0 1.00 1.00  0.87 0.9900031 0.9873008 1.00000 1.000000
#> 6  CYP2D6        <NA>   NA   NA   NA    NA        NA        NA 1.00000 1.000000
#> 7  CYP3A4   midazolam 12.5 0.57 0.96  1.00 0.9880276 0.9847995 1.62144 1.773674
#>        aucr  risk
#> 1        NA    NA
#> 2        NA    NA
#> 3 1.0106266 FALSE
#> 4 1.0129888 FALSE
#> 5 1.0111717 FALSE
#> 6        NA    NA
#> 7 0.4626871  TRUE