CYP perpetration risk table as per mechanistic-static modeling

CYP perpetration risk table as per mechanistic-static modeling

Usage

mech_stat_cyp_risk_table(
  perp,
  cyp_inh,
  cyp_ind,
  include_induction = T,
  substr = cyp_reference_substrates,
  na.rm = FALSE
)

Arguments

perp

The perpetrator object.

cyp_inh

CYP inhibiton data as data frame.

cyp_ind

CYP induction data as data frame.

include_induction

Boolean value to define whether induction should be included in the calculation (C-terms as per the FDA guideline)

substr

The CYP reference substrates to be used as data frame.

na.rm

Remove rows with lacking ki data (i.e., where ki == NA).

Value

A markdown-formatted table.

See also

Examples

mech_stat_cyp_risk_table(examplinib_parent, examplinib_cyp_inhibition_data, examplinib_cyp_induction_data)
#> 
#> 
#> Table: Mechanistic static modeling of CYP inhibition risk for  examplinib
#> 
#> |CYP     | $K_{i,u}$|substrate   | $F_{gut}$| $f_m$| $f_{m,CYP}$|$A_g$ |$A_h$ |$C_g$ |$C_h$ |AUCR  |risk  |
#> |:-------|---------:|:-----------|---------:|-----:|-----------:|:-----|:-----|:-----|:-----|:-----|:-----|
#> |CYP1A2  |        NA|tizanidine  |      1.00|  0.95|        0.98|NA    |NA    |1.00  |1.00  |NA    |NA    |
#> |CYP2B6  |        NA|NA          |        NA|    NA|          NA|NA    |NA    |1.00  |1.00  |NA    |NA    |
#> |CYP2C8  |      11.0|repaglinide |      1.00|  1.00|        0.61|0.986 |0.983 |1.00  |1.00  |1.011 |FALSE |
#> |CYP2C9  |      13.5|S-warfarin  |      1.00|  1.00|        0.91|0.989 |0.986 |1.00  |1.00  |1.013 |FALSE |
#> |CYP2C19 |      15.0|omeprazole  |      1.00|  1.00|        0.87|0.990 |0.987 |1.00  |1.00  |1.011 |FALSE |
#> |CYP2D6  |        NA|NA          |        NA|    NA|          NA|NA    |NA    |1.00  |1.00  |NA    |NA    |
#> |CYP3A4  |      12.5|midazolam   |      0.57|  0.96|        1.00|0.988 |0.985 |1.62  |1.77  |0.463 |TRUE  |