CYP perpetration risk table as per mechanistic-static modeling

CYP perpetration risk table as per mechanistic-static modeling

Usage

mech_stat_cyp_risk_table(
  perp,
  cyp_inh,
  cyp_ind,
  cyp_tdi = NULL,
  include_induction = TRUE,
  substr = cyp_reference_substrates,
  cyp_kdeg = cyp_turnover,
  na.rm = FALSE,
  show_dose = FALSE
)

Arguments

perp

The perpetrator object.

cyp_inh

CYP inhibition data as data frame. The following fields are expected:

'name' The name of the perpetrator compound.

'cyp' The CYP enzyme as (upper case) character.

'ki' The \(k_i\) in µM as numeric.

'source' Optional source information as character.

cyp_ind

The CYP induction data as data frame. The following fields are expected:

'name' The name of the perpetrator compound as character.

'cyp' The CYP enzyme as (upper case) character.

'emax' The \(E_{max}\), i.e., the maximum induction effect determined in vitro as numeric.

'ec50' The \(EC_{50}\) in µM as numeric.

'maxc' The maximal concentration in µM tested in the in vitro assay as numeric.

'source' Optional source information as character.

cyp_tdi

The CYP TDI data as data frame. The following fields are expected:

'name' The perpetrator compound name as character.

'cyp' The CYP enzyme as character.

'ki' The \(K_I\) in µM as numeric.

'kinact' The \(k_{inact}\) in 1/h as numeric.

'source' Optional source information as character,

include_induction

Switch to define whether induction effects should be included in the calculation (C-terms as per the FDA guideline)

substr

The CYP probe substrates to be used as data frame, defaults to cyp_reference_substrates. The data frame is expected to have the following fields:

'cyp' The CYP enzyme as (upper case) character.

'substrate' The substrate name as character.

'fgut' The fraction of the drug escaping gut metabolism.

'fm' The fraction of the drug that undergoes hepatic metabolism.

'fmcyp' The fraction metabolized by the respective CYP enzyme.

cyp_kdeg

The CYP turnover data as data frame. Defaults to the built-in reference data, cyp_turnover.

na.rm

Switch to define whether rows with lacking \(K_I\) or \(k_{deg}\) data are removed from the output.

show_dose

Show perpetrator dose in table title, defaults to FALSE.

Value

A markdown-formatted table.

See also

Examples

mech_stat_cyp_risk_table(
  examplinib_parent,
  examplinib_cyp_inhibition_data,
  examplinib_cyp_induction_data)
#> 
#> 
#> Table: Mechanistic static modeling of the CYP inhibition risk for examplinib
#> 
#> |CYP     | $K_{i,u}$|substrate   | $F_{gut}$| $f_m$| $f_{m,CYP}$|$A_g$ |$A_h$ |$B_g$ |$B_h$ |$C_g$ |$C_h$ |AUCR  |risk  |
#> |:-------|---------:|:-----------|---------:|-----:|-----------:|:-----|:-----|:-----|:-----|:-----|:-----|:-----|:-----|
#> |CYP1A2  |        NA|tizanidine  |      1.00|  0.95|        0.98|1.000 |1.000 |1.000 |1.000 |1.000 |1.000 |1.000 |FALSE |
#> |CYP2B6  |        NA|NA          |        NA|    NA|          NA|1.000 |1.000 |1.000 |1.000 |1.000 |1.000 |NA    |NA    |
#> |CYP2C8  |      11.0|repaglinide |      1.00|  1.00|        0.61|0.626 |0.983 |1.000 |1.000 |1.000 |1.000 |1.011 |FALSE |
#> |CYP2C9  |      13.5|S-warfarin  |      1.00|  1.00|        0.91|0.672 |0.986 |1.000 |1.000 |1.000 |1.000 |1.013 |FALSE |
#> |CYP2C19 |      15.0|omeprazole  |      1.00|  1.00|        0.87|0.695 |0.987 |1.000 |1.000 |1.000 |1.000 |1.011 |FALSE |
#> |CYP2D6  |        NA|NA          |        NA|    NA|          NA|1.000 |1.000 |1.000 |1.000 |1.000 |1.000 |NA    |NA    |
#> |CYP3A4  |      12.5|midazolam   |      0.57|  0.96|        1.00|0.655 |0.985 |1.000 |1.000 |6.885 |1.774 |0.232 |TRUE  |
mech_stat_cyp_risk_table(
  examplinib_compounds,
  examplinib_cyp_inhibition_data,
  examplinib_cyp_induction_data,
  examplinib_cyp_tdi_data)
#> 
#> 
#> Table: Mechanistic static modeling of the CYP inhibition risk for examplinib
#> 
#> |CYP     | $K_{i,u}$|substrate   | $F_{gut}$| $f_m$| $f_{m,CYP}$|$A_g$ |$A_h$ |$B_g$ |$B_h$ |$C_g$ |$C_h$ |AUCR  |risk  |
#> |:-------|---------:|:-----------|---------:|-----:|-----------:|:-----|:-----|:-----|:-----|:-----|:-----|:-----|:-----|
#> |CYP1A2  |        NA|tizanidine  |      1.00|  0.95|        0.98|1.000 |1.000 |1.000 |1.000 |1.000 |1.000 |1.000 |FALSE |
#> |CYP2B6  |        NA|NA          |        NA|    NA|          NA|1.000 |1.000 |1.000 |1.000 |1.000 |1.000 |NA    |NA    |
#> |CYP2C8  |      11.0|repaglinide |      1.00|  1.00|        0.61|0.626 |0.983 |1.000 |1.000 |1.000 |1.000 |1.011 |FALSE |
#> |CYP2C9  |      13.5|S-warfarin  |      1.00|  1.00|        0.91|0.672 |0.986 |1.000 |1.000 |1.000 |1.000 |1.013 |FALSE |
#> |CYP2C19 |      15.0|omeprazole  |      1.00|  1.00|        0.87|0.695 |0.987 |1.000 |1.000 |1.000 |1.000 |1.011 |FALSE |
#> |CYP2D6  |        NA|NA          |        NA|    NA|          NA|1.000 |1.000 |1.000 |1.000 |1.000 |1.000 |NA    |NA    |
#> |CYP3A4  |      12.5|midazolam   |      0.57|  0.96|        1.00|0.655 |0.985 |0.435 |0.476 |6.885 |1.774 |0.845 |FALSE |
#> 
#> 
#> Table: Mechanistic static modeling of the CYP inhibition risk for M1
#> 
#> |CYP    | $K_{i,u}$|substrate  | $F_{gut}$| $f_m$| $f_{m,CYP}$|$A_g$ |$A_h$ |$B_g$ |$B_h$ |$C_g$ |$C_h$ |AUCR |risk  |
#> |:------|---------:|:----------|---------:|-----:|-----------:|:-----|:-----|:-----|:-----|:-----|:-----|:----|:-----|
#> |CYP2C9 |       4.4|S-warfarin |         1|     1|        0.91|0.994 |0.994 |1.000 |1.000 |1.000 |1.000 |1.01 |FALSE |
mech_stat_cyp_risk_table(
  examplinib_compounds,
  examplinib_cyp_inhibition_data,
  examplinib_cyp_induction_data,
  examplinib_cyp_tdi_data, show_dose = TRUE)
#> 
#> 
#> Table: Mechanistic static modeling of the CYP inhibition risk for examplinib (450 mg)
#> 
#> |CYP     | $K_{i,u}$|substrate   | $F_{gut}$| $f_m$| $f_{m,CYP}$|$A_g$ |$A_h$ |$B_g$ |$B_h$ |$C_g$ |$C_h$ |AUCR  |risk  |
#> |:-------|---------:|:-----------|---------:|-----:|-----------:|:-----|:-----|:-----|:-----|:-----|:-----|:-----|:-----|
#> |CYP1A2  |        NA|tizanidine  |      1.00|  0.95|        0.98|1.000 |1.000 |1.000 |1.000 |1.000 |1.000 |1.000 |FALSE |
#> |CYP2B6  |        NA|NA          |        NA|    NA|          NA|1.000 |1.000 |1.000 |1.000 |1.000 |1.000 |NA    |NA    |
#> |CYP2C8  |      11.0|repaglinide |      1.00|  1.00|        0.61|0.626 |0.983 |1.000 |1.000 |1.000 |1.000 |1.011 |FALSE |
#> |CYP2C9  |      13.5|S-warfarin  |      1.00|  1.00|        0.91|0.672 |0.986 |1.000 |1.000 |1.000 |1.000 |1.013 |FALSE |
#> |CYP2C19 |      15.0|omeprazole  |      1.00|  1.00|        0.87|0.695 |0.987 |1.000 |1.000 |1.000 |1.000 |1.011 |FALSE |
#> |CYP2D6  |        NA|NA          |        NA|    NA|          NA|1.000 |1.000 |1.000 |1.000 |1.000 |1.000 |NA    |NA    |
#> |CYP3A4  |      12.5|midazolam   |      0.57|  0.96|        1.00|0.655 |0.985 |0.435 |0.476 |6.885 |1.774 |0.845 |FALSE |
#> 
#> 
#> Table: Mechanistic static modeling of the CYP inhibition risk for M1
#> 
#> |CYP    | $K_{i,u}$|substrate  | $F_{gut}$| $f_m$| $f_{m,CYP}$|$A_g$ |$A_h$ |$B_g$ |$B_h$ |$C_g$ |$C_h$ |AUCR |risk  |
#> |:------|---------:|:----------|---------:|-----:|-----------:|:-----|:-----|:-----|:-----|:-----|:-----|:----|:-----|
#> |CYP2C9 |       4.4|S-warfarin |         1|     1|        0.91|0.994 |0.994 |1.000 |1.000 |1.000 |1.000 |1.01 |FALSE |