Drug transporter inhibition risk

Drug transporter inhibition risk

Usage

transporter_inhibition_risk(
  perp,
  transporter_inh,
  transporter_ref = transporter_reference_data
)

Arguments

perp

The perpetrator object.

transporter_inh

Transporter inhibition data as data frame. The following fields are expected:

'name' The perpetrator compound name

'cyp' The UGT enzyme as (upper case) character.

'ic50' The \(IC_{50}\) of the inhibition in μM.

'source' Optional source information as character.

transporter_ref

Transporter reference data, see transporter_reference_data for details.

Value

A data frame.

Details

The relevant metric for the assessment of transporter interactions is \(R=[I]/K_i\). In in vitro transporter inhibition studies, \(IC_{50}\) values are experimentally determined. Since the transporter substrate concentration is usually kept very low in relation to \(K_m\) to minimize passive permeation, \(K_i = IC_{50}\) can be assumed.

The relevant perpetrator concentrations \([I]\) are:

\(I_{gut}\) for intestinal P-gp and BRCR

\(I_{max,inlet,u}\) for the hepatic basolateral transporters OCT1, OATP1B1 and OATP1B3

\(I_{max,ss,u}\) for the renal basolateral transporters OAT1, OAT3 and OCT2, as well as the apical transporters outside the intestinal mucosa, i.e., hepatic P-gp and BCRP, and MATE1, MATE2-k.

Note that the FDA and EMA guidelines differ in their definitions for the thresholds for assumed clinically relevant effects.

See also

Examples

transporter_inhibition_risk(examplinib_parent, examplinib_transporter_inhibition_data)
#>    transporter   ic50    source            r fda_thld fda_risk ema_thld
#> 1      Pgp_int   0.41 study 005 8.912391e+03     10.0     TRUE    10.00
#> 2      Pgp_sys   0.41 study 005 4.019984e-01      0.1     TRUE     0.02
#> 3     BCRP_int   1.90 study 005 1.923200e+03     10.0     TRUE    10.00
#> 4     BCRP_sys   1.90 study 005 8.674701e-02      0.1    FALSE     0.02
#> 5         OCT1   2.30 study 006 8.388626e-02       NA       NA     0.04
#> 6      OATP1B1 177.00 study 006 1.090047e-03      0.1    FALSE     0.04
#> 7      OATP1B3  35.00 study 006 5.512525e-03      0.1    FALSE     0.04
#> 8         OAT1 271.00           6.081894e-04      0.1    FALSE     0.04
#> 9         OAT3 300.00           5.493978e-04      0.1    FALSE     0.04
#> 10        BSEP  12.80           1.287651e-02      0.1    FALSE     0.02
#> 11        OCT2  67.00 study 006 2.459990e-03      0.1    FALSE     0.02
#> 12       MATE1   3.60 study 006 4.578315e-02      0.1    FALSE     0.02
#> 13      MATE2k   1.10 study 006 1.498358e-01      0.1     TRUE     0.02
#>    ema_risk
#> 1      TRUE
#> 2      TRUE
#> 3      TRUE
#> 4      TRUE
#> 5      TRUE
#> 6     FALSE
#> 7     FALSE
#> 8     FALSE
#> 9     FALSE
#> 10    FALSE
#> 11    FALSE
#> 12     TRUE
#> 13     TRUE